pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "vivipi"
version = "0.1.0"
description = "ViviPi scaffold for a Pico 2W-driven glanceable OLED monitor"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "PyYAML>=6.0.2",
]

[project.optional-dependencies]
dev = [
  "build>=1.2.2",
  "pytest>=8.3.5",
  "pytest-cov>=6.1.1",
  "ruff>=0.11.5",
]

[project.scripts]
vivipi-adb-service = "vivipi.services.adb_service:main"
vivipi-build-deploy = "vivipi.tooling.build_deploy:main"

[tool.hatch.build.targets.sdist]
include = [
  ".editorconfig",
  ".github",
  ".gitignore",
  "AGENTS.md",
  "build",
  "CLAUDE.md",
  "CODEX.md",
  "config",
  "docs",
  "firmware",
  "Makefile",
  "PLANS.md",
  "pyproject.toml",
  "README.md",
  "src/vivipi",
  "tests",
  "WORKLOG.md",
]

[tool.hatch.build.targets.wheel]
packages = ["src/vivipi"]

[tool.pytest.ini_options]
addopts = "--cov=src/vivipi --cov-branch --cov-report=term-missing --cov-report=xml --cov-fail-under=91"
testpaths = ["tests"]

[tool.coverage.report]
omit = [
  "firmware/*",
]
show_missing = true

[tool.ruff]
line-length = 100
target-version = "py312"
